requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
groq==0.11.0
TTS==0.22.0